Moved files to their correct places
