--- torch_geometric/contrib/nn/models/tokenGT/README.md
@@ -1,1 +1,0 @@
-Draft Implementation of tokenGT
