--- torch_geometric/contrib/nn/models/tokenGT/tests/README.md
@@ -1,1 +1,0 @@
-Test for tokenGT
